* issue-335:
  MobCost can exceed int, so calculate from long values
  Internal calculation use long then capped to Integer.MAX_VALUE

+++ src/main/java/ipsis/woot/power/calculation/UpgradePowerCalcDecapitate.java
@@ -14,7 +14,7 @@ public class UpgradePowerCalcDecapitate extends AbstractUpgradePowerCalc {
         if (!farmSetup.hasUpgrade(EnumFarmUpgrade.DECAPITATE))
             return;
 
-        int cost = spawnTicks * Woot.wootConfiguration.getInteger(
+        long cost = spawnTicks * Woot.wootConfiguration.getInteger(
                         farmSetup.getWootMobName(),
                         ConfigKeyHelper.getDecapPowerPerTick(farmSetup.getUpgradeLevel(EnumFarmUpgrade.DECAPITATE)));
         powerValues.upgradeCost += cost;
